backend/requirements.txt
# Django and REST Framework
Django==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# AI and ML Libraries
chromadb==0.4.15
sentence-transformers==2.2.2
openai==0.28.1
requests==2.31.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11

# Utilities
python-decouple==3.8
Pillow==10.0.1
numpy==1.24.3

# Development
python-dotenv==1.0.0

# Vector Database
faiss-cpu==1.7.4

# Text Processing
nltk==3.8.1
spacy==3.6.1

# HTTP and API
httpx==0.25.0
uvicorn==0.23.2

# File Handling
pathlib==1.0.1
os-sys==2.1.4

# Environment Management
python-environ==0.4.54

# Async Support
asgiref==3.7.2

# Security
cryptography==41.0.7

# Monitoring and Logging
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-django==4.7.0


